code review notes
